requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg[binary]
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv
aio-pika
pydantic-settings
colorlog
alembic
python-multipart
